+++ src/main/java/com/madisp/stupid/expr/MultiplicationExpression.java
@@ -26,7 +26,7 @@ public class MultiplicationExpression implements Expression {
         if (l instanceof Double || r instanceof Double) {
             return ctx.getConverter().toDouble(l) * ctx.getConverter().toDouble(r);
         }
-        return ctx.getConverter().toInt(l) * ctx.getConverter().toInt(r);
+        return ctx.getConverter().toLong(l) * ctx.getConverter().toLong(r);
     }
 
     @Override
